cleaned up repo
