+++ ExpectedCow.java
@@ -1,0 +1,16 @@
+package hw7;
+
+public class ExpectedCow {
+	public String expectedName() {
+		return "Cow";
+	}
+	
+	public String expectedFood() {
+		return "grass, hay, and corn";
+	}
+	
+	public String expectedInfo() {
+		return "milk and cheese";
+	}
+
+}
