--- bin/.gitignore
@@ -1,1 +1,0 @@
-/hw7/
